Chunk terrain modification saving !
Now chunk modifications are saved in a 00000101001010 fashion.
If a chunk is unmodified, on line 3 of it's .txt there's written x .
If it has ever been modified, the binary sequence of all the 256 tiles of the chunk (0 for not modified, 1 for modified) is written. Then these modifications are read upon chunk loading. Nicee