Slime and acid differentiation
Tweaked terrain generation for these biomes :
In slime biome, used periodic peaks functions (???) to create sort of repeated circular holes in the terrains, that give a sort of slimy feel to the terrain (it looks better than it's explained I swear). This is present all across the biomes, both in the slithery caves and bubble caves.
In acid biomes, using a seesaw function the terrain is really fucking weird with a lot of pixel holes (only in the bubble caves, not the slithery ones).