Obsidian biome differentiation
Obsidian biome generation has been modified !
Now in it, slithery caves are waaaay thinned and less variable, and bubble caves are not present, and replaced by periodic octogons (that are squashed and squished according to a perlin noise modifier).
Also, for this biome only, a smooth transition has been implemented, so it goes from hot biome to obsidiana biome smoothly (the octogons gradually appear and the slithery caves get thinner progressively... if all of the biome is near the humidity and temperature limits though, it might just look like a normal biome with no octogons due to the smooth transition. Hot and obsidian values have been changed, making them a bit more common and wide.