Smooth biome transitions
Biome transitions are now smooth !
Instead of having only one biome value, each pixel now has a list cointaining the "biome percentages" of each biome it has, that are sorted (from the biome with the higher percentage to the biome with the lowest percentage, i say percentage but it's an into from 1 to 100).
Then, the terrain of the tile is determined, by making a sort of "weighted mean" of the value you'd have with each biome if it was pure if that makes sense
Then the color of the biome is determined, then again using a weighted mean of all the colors of each biome that's in the biome list, and adding that to the base color (filled or not filled depending). It does find the color every frame anymore also, only at the start after loading the terrain, if a tile is dug or placed it then is updated.
Also, entities spawn according to the first biome in the list (ie, the dominant one), so if you got a 49% obsidian 51% hot it'll be a frog, if you have a 51% obs 49% hot it'll be an obsidian fairy etc...