STRUCTUUURES
Structures have been added !
On world gen, the game generates the structures on the surrounding megachunkgs when the player enters a new megachunk if they have not been generated already (a megachunk is 1024*1024 in size).
Then the generated structures are saved in the ChunkData as new chunk data terrain (warning I think creatures can't be generated cause of this).

Structures are generated, a random amount per megachunk (1 to 11 iirc), and for now can span like from 1*1 to 6*6 or smth, with both sizes able to vary. They contain empty squares with a filled 1 pixel thick border that can overlap (and just overlap without linking). The bigger the structure size, the more squares. For now it's only squares but more shit is to be adde nyehehehehhe. There is a new StructureData thing but for now it's useless