More structers ! Sawblades n stars
Added New Structures : Sawblades and stars.
The structure shapes are generated using functions over the distance to the center of the structure + the angle of the point to the center. This allows the length of tiles to be empty to vary depending on the angle of that pixel. Also, a filled outline is made to the shape by filling in the neighbouring tiles that have never been modified.