Water upgrade + Zoom
More water deplacement fixes, no weird chunk related water jump or teleportation or whatever.
Water is also not sand anymore, which is pretty nice. It will spread evenly when arriving in a cuve kinda idk. Not perfect but works (you can't have "piles" or "pyramids" of water anymore.
Also made a zoom/unzoom button, with S to zoom and D to unzoom (it just changes the amount of chunks rendered).

Currently working on water structures generation.