Waterlake BUGFIX
By changing the two values of the startingchunk to and Array with 2 numbers in it, the bug of loading has been FIXED ! The chunkStartingX changes were not kept on function exit  lol.
Now the lakes work nice! Miiight be a bug somewhere stilll but idk, they work fine now. Yee