Sprite Addition Begin Thing
Sprites.cs has been added ! Functions related to sprites and the sprite strings themselves will be put here.
Wrote a bit of the code for that, making OneSprite function, with a bitmap, palette, colorArray, dimensions, with funcitons to draw on a bigger bitmap (yet to do), draw the bitmap from the palette and color array (done), function to read the string into palette/colorarray/dimensions (yet to do).
I think that's all yea