Sprites ! PNG to string (palette, dimensions, colorarray), string (palette, dimensions, colorarray) to bitmap, copy and scaling into bigger bitmap...
Long title lmao.
A new .cs file has been created, sprites, that is dedicated to... sprites and other thing's management lol.
Functions for this : finding the sprites directory path, loading the png and parsing it into a newly created .txt file of the same name as the png, reading that .txt file and putting the data into an immediately useable bitmap...
Writing of a bigmap onto another bitmap (with scaling option)...

The format for the strings in the .txt :

The first line is the dimensions, it is exactly    "       x*y        ", with no : at the end

The second line is the palette.
Basically, every color in the image is separated into its 4 components, RGBA (in this order), and every color is written under the format r;g;b;a .
The 4 of these numbers are just one after another, so the string looks like a;b;f;q;f;g;e;w;f;a;e;g; but reading them 4 by 4 you can find the color.
Only unique colors are saved (including transparency), so if a sprite has 4 colors + transparency it will be like 20 numbers accounting for 5 unique colors in the second line.

The third line is every pixel in the image, one after the other, separated by ; (last character is a ; as well), in the format "x;" repeated, with x the index of the pixel on the palette (starting at 0).
If on second line the first 4 numbers are that one color transparent and the next 4 are that of color black, and the next 4 are that of color green, and the image is 3*3 dimensions, if on third line there is written :
0; 1; 0; 1; 2; 1; 0; 2; 0 then you'll have an original image that was :

010
121
020

and u replace the index of palette and u find it yea.
So I made a fishtest.png file, that was turned into a fishtest.txt by the program, that is succesfully loaded, and pasted onto the bigger bitmap at the bottom, in the right direction, with the right colors.

So yeeee it works :))))