The Spritening
A shitton of sprites added, plus dictionnaries for sprites and shit (based on index, had to do a dick because negative idx for liquids, might change that to like 2 arrays idk. One for entities as well (also a dict).
Fixed the blurry on resize bug for bitmaps (thanks stackoverflow guy)
Overlay shit looking like arrite for now yeeeehawwww