Inventory !
A working inventory (combinaison of an array for order and dictionnary for number) than can contain both entities and tiles (compounds).
As well, the current selected item AND the amount of that item that the player has appear on the right of the overlay. Yee.
If the value is -999, it is treated as infinite, and digging/placing don't change anything.
For every other object in the inventory, digging adds 1 to the count (impossible to dig entities for now), placing removes 1 to the count, and if the count is at 0 or lower the object gets removed from the inventory.
A cursor (int) is used to know which object is selected currently, and pressing C goes to previous object, V goes to the next one, and it loops.
Also, in Sprites.cs, a bitmap splicer function has been made, that splices an image into an array of bitmaps, using parameters to cut into columns and lines (x y i forgor the order).
The numbers are a sprite, from a big number sprite sheet that is sliced (0 to 9 plus infiniy symbol). Put in to a nubmer array with 11 bitmaps then.
Also there is not sprite for the "2" tile that was originally the player placed tile, uh, gotta make the sprite lol.
Goot bye.
The player can place entities like if they were generated by the game, but with a definite position and type (possible to spawn frog/fish in water, fish outta water, frog in fairy biome, fairy outside fairy biome...).
Color is random tho.
FindColor for entities, that found both color and type, has been split into FindColor and FindType.