The Honeying
it is NOW POSSIBLE for the HONRETS to make the HONEY ! ! ! !! ! HYAHOOO ! ! !

Now, entities can detect specific tiles of interest in plants and head towards it (only the flying ones fairies and hornets). Didn't put A* yet so they bump into walls on their path though lol.

AFter reaching a pollen tile on a plant when they're heading towards it, they now COLLECT it, and if they have more than 3 pollen in inventory they turn 3 pollen into a HONEY.

THen they can pour it on the ground.

Honey has a similar color to pee, but it is more orange (and tastes very different). Is is the liquid -5