THE A STARRING HELL YEAHHHHHHhhhhhhhhh
A STAR ALGORITHM WORKS I AM SO HAPPY YIPPPPPPPPPppppppppppppppeeeeeeeeeeeeeeeeeee yahoo woohoooooo yeaaahaeaaayyy bayyybeeeee

That's it. A* works, and the path is shown on rendering for debug purposes.
It can find the shortest (i hope) path to any 2 points on the map (if there is a path avaibable lol).
Path is clean and very pretty, pretty diagonal and angular but it's okay still ig.
No efffect for now, gonna have to actually make the path... following after lol it's gonna be a pain in the ass argh

Heuristic is not Djiklstra-ic, all nodes have equal lengths (except diagonal neighbors that have root of 2 approx, while horizontal/vertical neighbours have distance of 2).

WAIT A MINUTE THIS MEANS THAT IT ACTUALLY PATHFINDS IN DIAGONAL FUUCKKKKKKKKKKkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkk urgh my ghost tricks don't work i'm gonna have to make a smarter adding of vertices ergh.

Also added that when P button is help the game freezes lol