Better Pathfinding
Pathfinding (A*) has been updated !

Now, pathfinding CANNOT go through diagonals when there is no possibility to.
Like, to go from A to B:
XB
AX
it needs one of two of the Xs to be a tile that can be crossed (air, water idk).

Also, element is now taken into account when pathfinding (air has a cost of 1, while piss and water have
a cost of 3, and lava has a cost of 999999 so they can't pathfind through lava and avoid liquids).
For all entities it's the same tho I'll have to make a different one for different entities ig.

That's all xoxo bestie :heart: