The Path Simplifying
Simplify path function has been made !
The path is now simplified.
Amazing

From the returned diagonal A* path, whenever there is a straight orthogonal or diagonal segment, the inbetween of the extrement points of that segment are removed. Works nice ! The move a bit less stupid (it's not perfect tho still lmfaoooo).

Also debug to show the simplified path in red over the actual path in entity color