Nest Bugfix
Bugfixed nest generation, now all rooms are correctly linked to other with a corridor on generation (no rooms that are unlinked for no reason).

Added shape parameter for heuristic, to make nests with diamond or square roooms.
Works nice for diamond, less so for square. If shape is put at > 3 it can make for very nice + shaped rooms, interesting, will use that to make funky rooms maybe idk.

For now shape is always at root(2)/1.41xxxx