The actual NORNET diggENING !!! Hornests WORK (sorta)
HORNETS NOW DIG THEIR NESTS AND SHIT.

Instead of having the tiles turned to air when created, new corridors and rooms have all their tiles put in a DigErrand list.
Entities can now have an attached Nest reference, that points to their nest (duh), and the nest has a reference to all its entities.
Hornets now pathfind to a random DigErrand tile if reachable when they're idle or moving or smth, and just dig it like if they were digging a plant material (but they don't need to be on it now).
Works really nice, nests are now dug by hornets and that's super cool yeehaw !
Debug key shows the DigErrand queue in red for it to be more understandable.

what's left to do now is for hornets to put the HONEY they produce in the right place, make Nursery rooms, and reproduce with eggs and larva and shit. Should not be THAT hard. However :
The Loading/Unloading management of all diss shite needs to be done which is pain in da ass :(

also

Added a debug key (L or M i forgot lol) that toggles whethere entity paths and the DigErrands in nests are visible.
Added a fast forward key that speeds up shit up to 10 times faster (if computer iz not ded lol) (F is the key)