THE DEPOSITING
Hornets now DEPOSIT the HONEY in they're are ROOMZ !
Only type 2 rooms, that are open from the top, and they put it in there until -2y from the top of the room so it doesn't overflow. Works nice !
There is priority for depositing honey in these rooms in space available and hornet has honey in inventory, over gathering pollen in plants and digging.
That's all lol, basically like the digErrand list, but with tileWithFreeSpaceForHoney or sum shite
These tiles are in blue in the debug overlay, while digErrand are in red.
Also new drawpixel function for updateScreen shit makes it way easier to do shit ig