Entry corridor + noise generation changeeee
Added entrance corridors to nests ! Now, after generating the main room, the nest will generate a corridor that links this main room to an empty tile outside the nest, to link it to the outside so the hornets can like, go outside the nest lmfao. Works nice (but was a pain in the ass to make lmfao). Now I gotta do the saving in files shite :skull:

ALSO newnoise thing generation !
Now, like BigBiome, there is a new BigFill layer for cave Noise. Cave systems are "scaled" by *2, with a base of 64 tiles between points, but with a small layer at 25% strength of the big layer that works on a 16 tiles between points basis (like the original first version of the game), so now caves appear bigger than before, but have more like natural looking curves with like bumps and shit idk.

That's alllllllllllllll