Fog Of Warrrrrrrr
FOG OF WARRRRRRRRR has been added ! and it FUCKING WORKS

Basically, any tile that has not been yet explored/seen by the player is masked by a black color (including plants and entities present). the player character constantly casts raycasts 360° around it that remove that fog of war and allow vision of tiles. So now, unexplored areas are NOT visible and that's hella cool yeehaw !
Fog of war is saved in chunks as well, is the chunk is has not tile explored, both in game and in files, it has a value of 0 in exploration, if it's FULLY explored it has a value of 2 in exploration, and if it's parially explored it has a value of 1 in exploration AND an associated 32x32 bitmap and bool[32,32] that marks explored tiles, to decrease memory expanses and file size (yayy).
ITS WORKS SUPER NICE ! !  ! ! ! !

Also made hornets have a WAY smaller chance (1/2500 instead of 1/150, per frame) of stopping to move to a target location and becoming idlen should help with feeding and gathering and coming back to nest behaviors