Chunk loading and Saving of nests ! Also i'm back yayyyyyyyy
Big

Talk about :
- Megachunks
- Nest saving, orphan dict, whole nests being loaded and unloaded at the same time.
- New chunk loading system : faraway chunk, nest permanent load in megachunks till they get unloaded, the debug for loaded chunks...

Chunk loading and unloading have been changed COMPLETELY ! Now, since chunks are now in DICTIONARIES inside Screens, a new (smart...ish) unloading system has been made. Chunks are (mostly) loaded the same, when in range of the camera accounting for chunkResolution for maximum camera size and range of loading.
However, for unloading, they do not unload when out of range of the camera anymore : inside the Screen, the number of loaded chunks are counted : if that number is superior to chunkResolution^2 + the amount of Chunks that are deemed as permanently loaded, the game will test random chunks and sometimes unload them : a chunk can only be unloaded if : it's not a permanently loaded chunk (like a chunk inside of a nest), if it is farther from the camera's center position than 1.6 time the radius of chunk loading (chunkResolution*1.6), and if rand.next(x) says so lmao. The farthest a chunk is from the center, the more chance it has to be unloaded.
That should be all... BUT THEN NESTS EXIST D:
So...

Megachunks have been implemented ! Basically they are big fat chunk lookalikes that also get loaded and unloaded. They have a side that is of 512*512, so 16 chunks per side so 256 chunks in a megachunk. They basically store info related to structures and nests... as in the nest id and structure id in a list. WIth the id, since they're unique, in SctructureData and NestData the exact ones can be found.
Each frame, after chunks are loaded and unloaded, Megachunks are loaded unloaded : Megachunks who do not have any chunks in them loaded (chunks in nests DONT count) are unloaded, and megachunks that were not loaded but need to be cause they'd have chunks inside are loaded (phew).
When a megachunk is loaded, ALL nests inside are loaded and all the chunks inside the nests are loaded as well. When a megachunk is unloaded ALL the nests inside are unloaded. This allows for whole nests to be loaded/unloaded and not have nests partially loaded which would be kinda cringe ngl. It works fine.

When an entity that has a nestId is loaded, it  has the Nest instance reference attached to it if the Nest is loaded. If not, it is not added and the game adds it when the Nest is loaded afterwards.

Misc :
- Everything is now in dicts : activeEntities, activeNests, activePlants, activeChunks... instead of lists which is way easier to manage and way more flexible yay
- classes of Chunks, Screens, and Players, have been all put in separate files with their names.
- new SpriteStrings file that contain sprite strings to allow images not to be needed to create sprites back.
- Plant bugfix, + now ceiling plants can only be put on ceilings by the player and ground plants can only be put on the ground.
- Structures, like nests and entities and shit, now have unique ids that are stored in megachunk.

JFC this was HARD asf