LIGHT, new biomes, chandelier trees, and UNSAFE CODE >:)
LIGHT has been added in the game !

Basically works like fog of war (sorta). In biomes that are dark, the lightBitmap bitmap is fully WHITE on pixels that are supposed to be visible (enlightened), fully transparent on pixels that are supposed to be dar, and half transparent on those that are half illuminated etc...
This gets pasted on a layer the size of the gameBitmap. Then on this, circles of light of possible light sources (entities like fairies, leaves of chandelier trees, and player) white/transparentwhite bitmaps are copied (circles of different sizes and opacity of white depending on the light source, coming from a dict of possible sizes).
The, the bitmap gets inverted in transparency, all white pixels become transparent, and all transparent become black : makes the light come in effect : in every pixel that was not in a dark biome (or partially) and in every pixel in the range of light of something that emits light, there will be light. Yay.
(a darkBiome dict has been made to have all biome indexes that are dark lol)

In order to set pixels better, UNSAFE code with LockBits and shit is used... so it's faster.

Light radiuses (temp (or not)) :
Chanderlier Tree Light : 15 radius
Player : 11 light radius
Fairy : 7 light radius

NEW BIOMES !!!! CHandelier tree biome and Living biome
- Living is just Red lmfao (type = 10).
- Chandelier is gray, is DARK, and has vines and chandelier trees (new !) and frost fairies. (type = 9).
For testing, a Cold/Chandelier dimension type (type = 1, not monobiome) has been made.

CHANDERLIER TREES !!!!

Basically work like normal trees, except that they have rounded square branches (goes horizontally and then goes vertically with a little variable rounding at the angle).
Instead of leave bubbles, they have LANTERNS ! These are of that shape :
_ _X_ _
XXXXX
_ _O_ _
_XXX_
with X being wood (metal it's metal istg), and O being a leaf, that is yellow and is a LIGHT SOURCE for this type of tree : (1,1).

That's all. I think.