Colored lights and candles !
COLORED LIGHTS and CANDLES ??? yahoo !

Candles : type = 0, subtype = 1, so they are normal plant, they grow straight up with no random curvature, and have a flame made out of 2 colors to make it look like fire. ONLY CANDLES also are ANIMATED with 2 frames ! They have another bitmap in which another frame is stored, to make it look like the fire is going up and down.

The light system is now COLORED ! Was a pain in the ass to make ! How it works now :
LightBitmaps are now in a Dictionary not with an int=radius key, but with a (int, int, int, int)=(r,g,b,radius) key. r/g/b values from 0-255 get divided by 17 to turn them into a 0-15 range, and then lightBitmaps of specific r/g/b/radius values are made. Each time a light bitmap of a certain color and radius needs to be accessed, if not already in the dict (after color approximation), it is created. These bitmaps are thus colored, with transparency giving additional info, to ease pasting.

Instead of pasting with graphics each lightbitmap after positions of each light are taken, bytes for r, g, b are pasted onto the big lightbitmap one after another : for each channel r/g/b, if the new transparent*channel color of light is HIGHER than the one already in the byte, it is replaced (no merging of lights this way, and thus for r,g,b the highest value of light for each color is taken basically).
Then, the lightbitmap is pasted a bit differently on the gamebitmap :
Instead of pasting with transparency, for each pixel and each r/g/b, the gamebitmap pixel has acolor equal to that of its original non light color times the light value/255.

THUS COLORED LIGHTS WORK YEEEHAW !!!!

HOWEVER they are really FUCKING laggy ! Oh no. I will use threading to fix that HOPEFULLY

Misc : when spawning plants in chunks, the chunks tries to spawn 1 big plant (trees or chandelier), and 4 "small" type plants (basically all the other ones). So less cluttering with loads of trees/chandeliers but still lots of small plants