Animations !!!
OneAnimation class added ! Basically like a OneSprite but with an array.

Candles now don't have their secondary bitmap shit, but the animation that gets pasted directly on top of them, makes it easier to manage.

Entities/Plants/Players now have a lightColor for their light to make it easier.
LightColor is set to the mean of the color of the entity (or color of the thing that makes light for plants) with Color.White (so lights are less colored and tend towards white more to make them more blend in together and only have a faint color effect).

Yey