Biome Graphs, new findBiome method (laggy asf D: ), restructuration and undefined biome
Restructionration : functions moved to files where they fit more.
New Nest file also for nests (to separate from structures).

Implemented a method to make biome graphs depending on noise values (temperature, humidity, toxicity, acidity...). Since I can't see in 4d lmfao, 2 parameters are fixed and the 2 others are the ones changing along the 2 axises. Basically makes a bitmap, and for each point paints it the color of the tile that would be there if it was generated in the game.

New method to find biomes ! "heuristic" based sorta, instead of ranges (if temp < 500 etc...), now each existing biome is tested for it's distance to the ideal values present in the biomeTypicalValue dictionnary : for each of the 4 noise, the distance is measured, then all are added, and this gives scores : each score is substracted the max score - 100 to see if there one biomes dominates over the others (thus forcing one biome) or if the time is a mix of multiple (kinda like the old system with the transition zones).
It words fine !... EXCEPT IT'S SO FUCKING MORE EXPENSIVE... it actually takes SO much longer to compute which is very sad.

Undefined biome added ! It's just reaaally fucking pink/purple and it appears when there are no biomes in the biomeIndex array (when there's an issue LMAO)