LIQUID BUGFIX AFTER 9 MONTHS IT FUCKING WORKS YIPPEEEEEEeeeeeee
HOLY FUCKIN SHIT. I had written .Item1 instead of .Item2. On like 10 lines. Holy fuck I wanna off myself. But now it works. Holy shit.

Now, chunks with unstable liquid will be highlighted blue in debug mode, both on the "chunk debug minimap" and in game on the actual 32*32 chunk that there is creatures and plants in and terrain and u can go in there yes.

Also added a DICTIONNARY when doing the liquid push things, that remembers which uh tiles it already tested and that did not lead to any push so it stops there. Yoy. This reduces calculations by A LOT. Like /40I think it's super useful for water shit yayyyyy now water flows SUPER nice :D