Skeleton Fairiessss
Skeleton fairies ! THey're whitish-gray, more gray than frost fairies.
Are resistant to acid and lava ! So they are the only unkillable entity in the game since they are immune to fairy liquid too being already fairies lol.
When a SKELETON FISH is transformed by fairy liquid into a fairy, it doesn't become a normal fairy but a SKELETON FAIRY instead. That's the only way for them to spawn.