Tools n shite ! Pickaxe and Sword
ADDED TOOLS !!!
-They are the 5th type of item possible to be collected in inventories (x, x, 4)
-Pickaxe (1, 0) allows the player to dig. It must be selected in the inventory or player CANNOT DIG anymore.
-Sword (0, 0) allows the player to ATTACK and MURDER poor innocent creatures. That's so mean...
When equipped and the player presses X (tries to dig), it lauches a 5 step swing attack starting 90° from where the player is facing (if facing right, or diagonal right, the swing starts 90° clockwise from the actual player position, else if starts 90° anticlockwise from the player position, to make the swing go from bottom to top). Each frame the swing moves by 45°.
The swing is 2 pixels long (both on x/y axis and diagonal), and has 8 possible angles. It is drawn as 2 white pixels, but the hitbox is bigger. For diagonal positions, it is the 2*2 square that contains the sword. For axis positions, it is the 2 sword pixels + the 4 pixels on the sides of these 2 pixels (so a 3*2 or 2*3 rectangle). All entities that are in these positions take 1 damage, except if they have already been hit by that exact same swing during a previous frame.

-Entities now have a certain amount of hp (depending on their type), that decreases when hit.
-When entities get hit (and don't die) they become bright red for a short amount of time before regaining their normal color over 0.5s.
-Player now has a direction that updates when moving (x/y axis + its 4 diagonals) based on input. It can't not face a direction. Priority to x axis direction when player holds both x and y and stops holding them at the same time (will have x orientation and no y orientation). Used for attacks (only... for nowHUE HUE HUE).

-Added sprites for hornet eggs, larvae, and cocoons (and sword and pickaxe)