BLOOD ALTAR WORKING !!!! No portal yet but THAT'S COMING SOON DONT BE IMPATIENT
BLOOD ALTAR LEGIT JUST FUCKING WORKS WTF ????

--- MAGIC WAND ---

Player can now "attack" with the magic wand. However it does not affect creatures at all. 1 frame of attack, brown pixel on top of the player. 2 frame, brown pixel up diagonal of the player. Frame 3 to 6, brown pixel on the side of the player the attack is facing, and PURPLE pixel 3 tiles away horizontally from where the player is facing. This exact purple pixel (and it only) counts as the attack.
THE ATTACK is a special one. It doesn't hurt creatures, but it will instead TEST FOR BLOOD ALTAR on the tile it's on. This allows to trigger the blood altars the player has built, and have the test be a trigger so the game only needs to check when the player uses the wand lmao else it'd be a mess.
(for now, if the altar is correctly a blood altar, the flesh tile just turns into an acid tile lol, but there's gonna be a portal to the flesh dimension at some point instead I swear lol).

--- BLOOD ALTAR ---

To be valid as a blood altar, a structure (made by player) must :
-have a 1 tile tall pool of blood, made of between 3 to 15 (both included) of blood tiles
-Each tile of blood must have a tile of Dense Rock under it and a tile of air (so nothing) over it
-on each edge of the pool of blood, there must be a tile of Dense Rock, on the same level as the blood tiles.
This looks like this (X is dense Rock, B is blood, _ is air, = can be any tile) :
=__________=
=XBBBBBX=
==XXXXX==
Now to TRIGGER the blood altar, the player must place a tile of FLESH over the pool of blood ( not directly 1 tile over it or it won't work). That tile needs to be between 2 and 5 (both included) tiles over the y position of the blood pool, and it needs to be on the same x coordinate as a blood tile of the blood pool. It also needs to have its 8 surroundings tiles be air (nothing).
After all the conditions are met, the player must then use their MAGIC WAND that they have crafted to "enmagicken" idk the tile of flesh. When the purple magic pixel appears on the flesh tile... the portal will appearWELL NO FUCK YOU. For now the flesh tile just turns into an acid pixel. Fuck you I didn't make the rest of the code. It will come soon tho. BUT IT FUCKING WORKS FOR NOW YEEEHAAWWWWWWWWwwwwwwwwww

--- MISC ---

-Code cleanup and shit
-New recipe : 3 fleshTile to 1 blood (temp but might keep)
-Byebye ObsidianFairy.alpha.png. I will miss you. Unironically.