Mold Partikelz
-MOLD PLANTS NOW SPAWN PARTICLES. Spawn from a random tile in the plant's fillDict, 1/10 chance per frame.
type : (5, x, y). x and y are type and subtype of the material in the plant colorDict to take the color from.
new refId variable that has the plant that spawned it's ID (base value -1) to be able to retrieve color.
(5, x, x) particle are Angular, have random angle at start, and move randomly, changing angle and speed over time. They are of the color of the mold material in color dict of the mold that spawned them.

-Misc : added a devMode bool, when set to false, inventory is just put to the 3 base tools (pick, sword, scythe), when set to true, it's the full debug inventory with infinite stuff.