NEW CHUNKGETTING SYSTEM and bugfixs n shite
--- NEW SYSTEM OF CHUNKGETTING/LOADING ! ---
-Basically like the MegaChunk getting. Promoting from extraLoaded to LoadedChunks, and also demoting when unloading (when unloaded Chunks are not byebye immediately, they're put in extraloaded, it takes 5 frames after they've been put there for them to be unloaded, timer of an ExtraLoaded chunk reinitialises everytime the game gets it).
---------------> A SHITTON of changes in Chunks and how they're loaded. Liek wtf.
Chunk generation has been sperated into lots of functions instead of 1 big block. First determinecontents (biome, then terrain if chunk is loaded for the first time). New creation of Chunk instances from Json as parameters (like structures, entities, plants...).
Promotion from extra to loaded : called when extra to loaded (duh) but also when loading a chunk fully when it's not in the extra. Takes care of loading entities and plants, generating them if they haven't, loading fog of war stuff and loading plants/entities from Jsons if there is any (only case where there is none passed is when chunk is loaded for the first time and fully loaded at the same time, else either takes the json used when fully loading during the first part of chunk making, or loads the Json as it can have been modified when promoting from extra).

-> Player is now placed again in an empty tile when spawning for the first time ! Doesn't do it on debug for debugging purposes. Probably cannot be put in ocean biome because of this (yay) but I'll have to check.
A spiral checking function was used for that, that can be reused easily (get next pos in the clockwise spiral starting upwards centered on 0 from a pos).

- Fixed a bug with lake sizes being always set to 2000 max... and not from 1-11008, bruh. Now it works tho so there are gonna be megalakessssssss. Had to decrease max size due to stackoverflow >:( but prolly can increase it again after structure generation is optimized not to take place at the same time for each megaChunk (but spread over multiple frames to make it go smoother).

-LOADS of code cleanup thanks to that and optimizations :D :D

PHEW ffs this took a shitton of time