Smooooth zoomeratoooooor
Game is now SMOOTH no matter the zoom level !
PNGmultiplicator is now a value of Game class : every time zoom/dezoom happens (and on startup), it is chosen depending on the zoom level : Max(1, 320/zoomlevel). Thus, the zoomiest the game is, the higher the PNGmultiplicator, and there is no blur anymore even when only 4-5 pixels of resolution (since it's been upscaled accordingly). This also removed the weird white border bug. Fuuuck that's way better

-> Left to do : scaling/repositioning of the debug loaded chunks/megachunks minimaps because it's really the bordel for now