Better zooming and fixed chunk debug minimap ù!
Yey ! New way of changing the zoomLevel, using a zooming speed. Works very very nice ! Unzooming/zooming becomes slower in terms of pixel at the zoomLevel decreases (meaning the screen is more zoomed in). So when being close, it unzooms slowly, the farthest away, the fastest it becomes. yey. Use of a float realZoomLevel for that, and float zoomingSpeed.

Also fixed the minimap by modifying the drawPixelFixed and adding 3564354 unecessary arguments to it (but it works... fine. enoug)

yerg