New movement shite
Entity and Player movement is now the SAME
-> Movement system changed to make it better (lol). Before, if would move on the Y axis, stop (if bumping or moved far enough), then do the same on the X axis.
NOW, it moves similarly to a raycast from pos to target pos, except that is is able to slide. If for example it bumps on a tile on x axis, it will not set speedX = 0 and stop moving on x axis, but try moving on y axis. If it suceeds, it will try moving on x axis again. Only stops movement and sets speed to 0 on an axis, if blocked on the 2 axises, or if blocked in that axis and done moving on the other axis.
Normal movement like a rayCast, (for ex if going from (0, 0) to (6, 3), it will to right right up right right up right right up. But allows deviations. If blocked on the right for ex, it can go right up up right righ right up. Allows for smoother movement and not blocking entities/player most importantly.
Works very fine !

Also same function for Entities and Player, as its in Entity class. Saving in chunks when exiting carried out by a small virtual overrided function (for it not to be carried out for players). I didn't test it tho but it shouuuuld work (function ran when getting Chunk fails)