EntityBehaviorssszzzzzzz
Instead of all the fucking dictionaries for swimming, digging, the drops, the starting hp...... of entities, ENTITYBEHAVIORS have been introduced ! (int type, int subType) -> EntityBehavior (for the dict)
Basically data holder classes, instantiated on startup for each type of existing entity :
Contains for each entity, its name, if it can fly, dig (as in worm), swim, and skip on water. Also contains the drops, the starting HP, and the color of entity of that type.
COLOR is separated in 3 color variables (one for r, one for g, and one for b). Each one of these 3 color variables is a tuple (v, h, s) :
v : the base color value of that entity, for that color channel
h :  the HUE modifier.
s : the SHADE modifier
When finding out the color of the entity, for each channel, it will create a color with the base color, + the value in the hue, + the value in the shade. Possible to put negative for hue/shade (so that hue shifts from red to green for example idk). When making the color, the same shade variation is used in the 3 channels (based on seed), multiplied by the modifier in the EntityClass
so if shade for r g b is 20 20 30, and that entity has -0.5f as shadeModifier, it will put -10 -10 -15 on top of the base color.

PHEW.

It works ! Colors have changed a tiny bit... but eh idgaf