Bridges prevention in plants (i mean gaps)
plantGrowthRules now contain a preventBridges bool. If set to true, it will prevent bridges when the plant grows, as simple as that (it will cap the difference between lastDrasPos and new DrawPos at +1 or -1, for both axes independantly, so that it doesn't jump).
Tested with many fleshTree2, seems to workkkkkkkkkkkkkk
-> by default preventBridges is true, but can be set to false if u want ig

i just realised "BridgesPrevention" makes no fucking sens forget eveything it's replaced to GAP PREVENTION LMFAOOOOO NOT BRIDGES