Fish WingTails lol
WingTraits in entity can now be used for fish tail/fins !!

Basically wingTraits now contains a POS and a TYPE (and color is now isVariation/argb tuple like in tailMap). Pos will be used for long fish to put the fin places

When type is 0 : normal wings
When type is 1 : fish tail fin.
--> based on movement speedX/Y, 3 positions uses rotation and period. If fish moves top right it will be located bottom left (+-45°). So it adapts to fish direction to be its tail. And if fish immobile (or super tiny movement) it's straight up not present lol.
---> Works super fine ! For player also.