Maturation TFTts and new chunk maturity system !!!
Maturation TFTs !!!!!! An evolution of the surfaceMaterial system, in which the Chunk was matured before being fully loaded to prevente infinite chunk generation (as chunk maturation needs to take into account the surrounding tiles, even in never generated chunks, which causes infinite recursion).
--> Chunks now have 3 LEVELS OF MATURATION !!!
-Level 0 : Terrain Generated, with base TFTs applied
-Level 1 : Terrain Generated, with base TFTs applied, and maturation TFTs applied
-Level 2 : Terrain Generated, with base TFTs applied, maturation TFTs applied, and Plants + Entities generated
(that maturation Level is ofc stored in ChunkJSONs)

By default, any getting of chunk/tile will force the Chunk it's getting it in to be generated to Level 1 : Terrain, and all TFTs. So that the terrain is FULLY made before stuff like stuctures, etc, are generated.
However, some TFTs require the surrounding tiles to be used : like the "Skin" TFT, previously in surfaceMaterial, that needs to be exposed to at least 1 air tile to become skin. These TFTs will get tiles from chunks but only forcing them to be generated up to Level 0, so that it doesn't cause any infinite recusion !!!! Works very fine, basically merging of the previous system with the TFT system.
---> So now, TFTs have a isMaturation bool, and are automatically stored in BiomeTraits in 2 different TFT arrays.
------------> Promotion to fullyLoaded forces maturation to Level 2 (fully generated with plants n shit).

-> In the chunk debug minimap, Maturation Level 1 chunks are whitish instead of the light green of Maturation Level 2 chunks (both being in Extraloaded).

-> A temporary ManyValues class is used to store values after maturation Level 0 generation when it's getting generated to Level 1 (so like amost always). Very weird solution and very temporary but it works LMAO. Since values need to b generated again for maturation TFTs. It gets set to null to not waste memory after maturation to Level 2. I think after testing it actually makes it a lot faster lol but idk.

--> Next I'll do dirt and mud TFTs with that yeyeyeye !!!
IDEA : Make persistance of ManyValues for 8 directional neighbours ? For crystals that will b updated l8r ??? idk